:person_fencing: Removing need for init/deinit calls.
We can just initialize them the first time we access that data with atomics to make it thread-safe.

Especially since you don't need to call init/deinit if you're not using ORIGINAL_STRINGS, I'd rather those not be there if we can avoid it. Also makes usage of the module that much more simple.