:crayon: Making ORIGINAL_STRINGS initialization more robust
Fixing potential race condition with initialization of the hash-to-string table with the values saved at compile-time. It could have been the case that multiple threads try to do the initialization, only one would perform it but then the other would go onto grab the lock before the initializing thread and start accessing the hash-to-string table while it was still beign filled by the initializing thread.

Now any threads that potentially race to initialize wait on the one that actually got to initialize things.