:bat: Updating to beta version 0.2.017
Replacing calls to `table_find` to `table_find_new` since the API is changing to returning the bool to say that the element was found first, then the element itself. We'll need another change to rename all the `table_find_new` to be `table_find` later when the old API is finally removed.